Add an initial FedoraWiki.login method.
It does not work yet, but it's a start.  We can easily inject the session
cookie (which may or may not be named properly), but I'm thinking that we need
to also inject the username cookie in there as well.  If so, we will need to
add support for injecting arbitrary cookies into the ProxyClient.