Use a generic openid_insecure instead of a specific one
Add a generic openid_insecure that is valid for the whole object instead of relying on an
`insecure` argument passed on the login() method.
This is similar to the `insecure` attribute set in the constructor of the object to allow
querying applications with self-signed certificates.

By this mechanism we can test an application in stg using the OpenID provider in prod or
vice-versa.