Return in json all the arguments to forward to the openid server if asked
If the user asks for json return him/her all the argument to pass to the
openid server to create the session. This will allow us to create openid
CL client without having to parse the HTML to retrieve the arguments to
give to the openid server.

Fixes https://github.com/fedora-infra/python-fedora/issues/93