Update the OpenidProxyClient to rely on the JSON output of flask_fas_openid
flask_fas_openid if asked will return the argument to send to the openid
server in JSON. This mean no more HTML scrapping to retrieve the info.

This commits drop the dependency on BeautifulSoup4 to parse the html form
to retrieve the arguments to send to the OpenID server directly by relying
on the JSON output.