fas2: Drop the override on my email address
A while ago, I wanted to use my @fp.o alias for my RHBZ account.

I haven't been doing that for a while now, and there's apparently a
hotfix on pkgdb that removes these 2 lines.

This commit removes the 2 lines, so that:

1. the hotfix is not necessary any more
2. I stop receiving emails every time there's an update and the hotfix
   hasn't been applied yet

(Also, wow, that gmail address has been closed for so long!)