openidbaseclient: clear cookies on login
If we previously had a session that is now expired, or became invalid for
any other reason, we should make sure to clear those cookies before we
issue the login.
The problem is that Flask, when presented with an invalid session, first
sends a new, blank session, and that does not overwrite the previously
existing cookie in requests, but it does take precedence.
That means that if we later get a session with a return URL, requests
will still send the cleared cookie, breaking the login process.

Fixes: RHBZ#1317579

Signed-off-by: Patrick Uiterwijk <puiterwijk@redhat.com>