Retire the vast majority of code
This drops the code for the FAS auth modules for:
- TurboGears1
- TurboGears2
- Django

All applications should have moved over to OpenID (Connect) by now.

This drops the FAS client, FAS is going away for Fedora rather soon.

This drops the OpenID Base clients: for API authentication, applications
need to switch to OpenID Connect.
At ths moment, Bodhi is the one client I am aware of that uses this,
and it will be moved to OIDC soon.

Signed-off-by: Patrick Uiterwijk <patrick@puiterwijk.org>